requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
schedule==1.2.0
googletrans==4.0.0rc1
openai==1.3.0
feedparser==6.0.10
